Solution to Find the Maximum and Minimum Element in an Array

+++ Find the Maximum and Minimum Element in an Array@@676c02002dbc8c3e7a280b70/code.c
@@ -7,8 +7,8 @@ int main(){
         scanf("%d",&arr[i]);
     }
     for(int i=0;i<n;i++){
-        int min=arr[i]%i>=1;
-        int max=arr[i]%i<=1;
+        float min=arr[i]%i>=1;
+        float max=arr[i]%i<=1;
         printf("%d %d",min,max);
     }
 }
